Versao 1.0.2 - Foreign Keys Adicionadas

+++ src/main/java/com/pooflix/poo/repositorie/ActorDto.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.pooflix.poo.models.Actor;
 
-public interface ActorDto extends CrudRepository<Actor, Long> {
+public interface ActorDto extends CrudRepository<Actor, Integer> {
 
 }
